Move ScanConfig tests into the new structure
