+++ tests/protocols/gmpv208/entities/scan_configs/test_create_scan_config.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -19,50 +19,40 @@
 from gvm.errors import RequiredArgument
 
 
-class GmpCreateConfigFromOSPScannerTestCase:
-    def test_create_scan_config_from_osp_scanner(self):
-        self.gmp.create_scan_config_from_osp_scanner('a1', 'foo')
+class GmpCreateScanConfigTestMixin:
+    def test_create_scan_config(self):
+        self.gmp.create_scan_config('a1', 'foo')
 
         self.connection.send.has_been_called_with(
             '<create_config>'
-            '<scanner>a1</scanner>'
+            '<copy>a1</copy>'
             '<name>foo</name>'
             '<usage_type>scan</usage_type>'
             '</create_config>'
         )
 
-    def test_create_scan_config_from_osp_scanner_with_comment(self):
-        self.gmp.create_scan_config_from_osp_scanner(
-            'a1', 'foo', comment='comment'
-        )
+    def test_create_scan_config_with_comment(self):
+        self.gmp.create_scan_config('a1', 'foo', comment='comment')
 
         self.connection.send.has_been_called_with(
             '<create_config>'
             '<comment>comment</comment>'
-            '<scanner>a1</scanner>'
+            '<copy>a1</copy>'
             '<name>foo</name>'
             '<usage_type>scan</usage_type>'
             '</create_config>'
         )
 
-    def test_create_scan_config_from_osp_scanner_missing_scanner_id(self):
+    def test_create_scan_config_missing_scan_config_id(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.create_scan_config_from_osp_scanner(
-                scanner_id='', name='foo'
-            )
+            self.gmp.create_scan_config(config_id='', name='foo')
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.create_scan_config_from_osp_scanner(
-                scanner_id=None, name='foo'
-            )
+            self.gmp.create_scan_config(config_id=None, name='foo')
 
-    def test_create_scan_config_from_osp_scanner_missing_name(self):
+    def test_create_scan_config_missing_name(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.create_scan_config_from_osp_scanner(
-                scanner_id='c1', name=None
-            )
+            self.gmp.create_scan_config(config_id='c1', name=None)
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.create_scan_config_from_osp_scanner(
-                scanner_id='c1', name=''
-            )
+            self.gmp.create_scan_config(config_id='c1', name='')
