+++ tests/protocols/gmpv208/entities/scan_configs/test_modify_scan_config_set_comment.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2020-2021 Greenbone Networks GmbH
+# Copyright (C) 2018-2021 Greenbone Networks GmbH
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 #
@@ -19,32 +19,36 @@
 from gvm.errors import RequiredArgument
 
 
-class GmpModifyConfigSetNameTestCase:
-    def test_modify_scan_config_set_name(self):
-        self.gmp.modify_scan_config_set_name('c1', 'foo')
+class GmpModifyScanConfigSetCommentTestMixin:
+    def test_modify_scan_config_set_comment(self):
+        self.gmp.modify_scan_config_set_comment('c1')
 
         self.connection.send.has_been_called_with(
             '<modify_config config_id="c1">'
-            '<name>foo</name>'
+            '<comment></comment>'
             '</modify_config>'
         )
 
-    def test_modify_scan_config_set_name_missing_config_id(self):
+        self.gmp.modify_scan_config_set_comment('c1', 'foo')
+
+        self.connection.send.has_been_called_with(
+            '<modify_config config_id="c1">'
+            '<comment>foo</comment>'
+            '</modify_config>'
+        )
+
+        self.gmp.modify_scan_config_set_comment('c1', comment=None)
+
+        self.connection.send.has_been_called_with(
+            '<modify_config config_id="c1">' '<comment/>' '</modify_config>'
+        )
+
+    def test_modify_scan_config_set_comment_missing_config_id(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name(config_id=None, name='name')
+            self.gmp.modify_scan_config_set_comment(config_id=None)
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name('', name='name')
+            self.gmp.modify_scan_config_set_comment('')
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name(config_id='', name='name')
-
-    def test_modify_scan_config_set_name_missing_name(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name(config_id='c', name='')
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name(config_id='c', name=None)
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_name('c', '')
+            self.gmp.modify_scan_config_set_comment(config_id='')
