+++ tests/protocols/gmpv208/entities/scan_configs/test_modify_scan_config_set_scanner_preference.py
@@ -19,89 +19,70 @@
 from gvm.errors import RequiredArgument
 
 
-class GmpModifyConfigSetNvtPreferenceTestCase:
-    def test_modify_scan_config_set_nvt_pref(self):
-        self.gmp.modify_scan_config_set_nvt_preference(
-            config_id='c1', nvt_oid='o1', name='foo'
+class GmpModifyScanConfigSetScannerPreferenceTestMixin:
+    def test_modify_scan_config_set_scanner_pref(self):
+        self.gmp.modify_scan_config_set_scanner_preference(
+            config_id='c1', name='foo'
         )
 
         self.connection.send.has_been_called_with(
             '<modify_config config_id="c1">'
             '<preference>'
-            '<nvt oid="o1"/>'
             '<name>foo</name>'
             '</preference>'
             '</modify_config>'
         )
 
-        self.gmp.modify_scan_config_set_nvt_preference('c1', 'foo', 'o1')
+        self.gmp.modify_scan_config_set_scanner_preference('c1', 'foo')
 
         self.connection.send.has_been_called_with(
             '<modify_config config_id="c1">'
             '<preference>'
-            '<nvt oid="o1"/>'
             '<name>foo</name>'
             '</preference>'
             '</modify_config>'
         )
 
-    def test_modify_scan_config_set_nvt_pref_with_value(self):
-        self.gmp.modify_scan_config_set_nvt_preference(
-            'c1', 'foo', nvt_oid='o1', value='bar'
+    def test_modify_scan_config_set_scanner_pref_with_value(self):
+        self.gmp.modify_scan_config_set_scanner_preference(
+            'c1', 'foo', value='bar'
         )
 
         self.connection.send.has_been_called_with(
             '<modify_config config_id="c1">'
             '<preference>'
-            '<nvt oid="o1"/>'
             '<name>foo</name>'
             '<value>YmFy</value>'
             '</preference>'
             '</modify_config>'
         )
 
-    def test_modify_scan_config_set_nvt_pref_missing_nvt_oid(self):
+    def test_modify_scan_config_scanner_pref_missing_name(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', 'foo', nvt_oid=None, value='bar'
+            self.gmp.modify_scan_config_set_scanner_preference(
+                'c1', name=None, value='bar'
             )
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', 'foo', nvt_oid='', value='bar'
+            self.gmp.modify_scan_config_set_scanner_preference(
+                'c1', name='', value='bar'
             )
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', 'foo', '', value='bar'
-            )
-
-    def test_modify_scan_config_nvt_pref_missing_name(self):
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', name=None, nvt_oid='o1', value='bar'
-            )
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', name='', nvt_oid='o1', value='bar'
-            )
-
-        with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                'c1', '', nvt_oid='o1', value='bar'
+            self.gmp.modify_scan_config_set_scanner_preference(
+                'c1', '', value='bar'
             )
 
     def test_modify_scan_config_set_comment_missing_config_id(self):
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                config_id=None, name='foo', nvt_oid='o1'
+            self.gmp.modify_scan_config_set_scanner_preference(
+                config_id=None, name='foo'
             )
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference('', 'foo', 'o1')
+            self.gmp.modify_scan_config_set_scanner_preference('', 'foo')
 
         with self.assertRaises(RequiredArgument):
-            self.gmp.modify_scan_config_set_nvt_preference(
-                config_id='', name='foo', nvt_oid='o1'
+            self.gmp.modify_scan_config_set_scanner_preference(
+                config_id='', name='foo'
             )
