+++ tests/protocols/gmpv208/entities/scan_configs/test_sync_scan_config.py
@@ -16,27 +16,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gvm.errors import GvmError
 
+class GmpSyncScanConfigTestMixin:
+    def test_sync_scan_config(self):
+        self.gmp.sync_scan_config()
 
-class GmpDeleteConfigTestCase:
-    def test_delete_scan_config(self):
-        self.gmp.delete_scan_config('a1')
-
-        self.connection.send.has_been_called_with(
-            '<delete_config config_id="a1" ultimate="0"/>'
-        )
-
-    def test_delete_scan_config_ultimate(self):
-        self.gmp.delete_scan_config('a1', ultimate=True)
-
-        self.connection.send.has_been_called_with(
-            '<delete_config config_id="a1" ultimate="1"/>'
-        )
-
-    def test_delete_scan_config_missing_scan_config_id(self):
-        with self.assertRaises(GvmError):
-            self.gmp.delete_scan_config(None)
-
-        with self.assertRaises(GvmError):
-            self.gmp.delete_scan_config('')
+        self.connection.send.has_been_called_with('<sync_config/>')
