+++ tests/protocols/gmpv214/test_v208_from_gmpv214.py
@@ -62,22 +62,8 @@ class Gmpv214GetFeedTestCase(GmpGetFeedTestCase, Gmpv214TestCase):
     pass
 
 
-class Gmpv214CloneConfigTestCase(GmpCloneConfigTestCase, Gmpv214TestCase):
-    pass
-
-
 class Gmpv214ClonePolicyTestCase(GmpClonePolicyTestCase, Gmpv214TestCase):
 
-    pass
-
-
-class Gmpv214CreateConfigTestCase(GmpCreateConfigTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214CreateConfigFromOSPScannerTestCase(
-    GmpCreateConfigFromOSPScannerTestCase, Gmpv214TestCase
-):
     pass
 
 
@@ -91,20 +77,8 @@ class Gmpv214CreatePolicyTestCase(GmpCreatePolicyTestCase, Gmpv214TestCase):
     pass
 
 
-class Gmpv214DeleteConfigTestCase(GmpDeleteConfigTestCase, Gmpv214TestCase):
-    pass
-
-
 class Gmpv214DeletePolicyTestCase(GmpDeletePolicyTestCase, Gmpv214TestCase):
 
-    pass
-
-
-class Gmpv214GetConfigTestCase(GmpGetConfigTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214GetConfigsTestCase(GmpGetConfigsTestCase, Gmpv214TestCase):
     pass
 
 
@@ -440,10 +414,6 @@ class Gmpv214GetVersionCommandTestCase(
     pass
 
 
-class Gmpv214ImportConfigTestCase(GmpImportConfigTestCase, Gmpv214TestCase):
-    pass
-
-
 class Gmpv214ImportReportFormatTestCase(
     GmpImportReportFormatTestCase, Gmpv214TestCase
 ):
@@ -451,40 +421,6 @@ class Gmpv214ImportReportFormatTestCase(
 
 
 class Gmpv214ModifyAuthTestCase(GmpModifyAuthTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214ModifyConfigTestCase(GmpModifyConfigTestCase, Gmpv214TestCase):
-    pass
-
-
-class Gmpv214ModifyConfigSetCommentTestCase(
-    GmpModifyConfigSetCommentTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214ModifyConfigSetFamilySelectionTestCase(
-    GmpModifyConfigSetFamilySelectionTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214ModifyConfigSetNvtPreferenceTestCase(
-    GmpModifyConfigSetNvtPreferenceTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214ModifyConfigSetNvtSelectionTestCase(
-    GmpModifyConfigSetNvtSelectionTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214ModifyConfigSetScannerPreferenceTestCase(
-    GmpModifyConfigSetScannerPreferenceTestCase, Gmpv214TestCase
-):
     pass
 
 
@@ -506,12 +442,6 @@ class Gmpv214RestoreTestCase(GmpRestoreTestCase, Gmpv214TestCase):
 
 class Gmpv214SyncCertCommandTestCase(
     GmpSyncCertCommandTestCase, Gmpv214TestCase
-):
-    pass
-
-
-class Gmpv214SyncConfigCommandTestCase(
-    GmpSyncConfigCommandTestCase, Gmpv214TestCase
 ):
     pass
